refactor : 텍스트필드 placeholder, size 수정 , 타입 파일 제거
