--- src/shared/ui/textfield/Textfield.types.ts
@@ -1,8 +1,0 @@
-export type TextFieldSize = 'sm' | 'md' | 'lg';
-
-export interface TextFieldProps {
-  placeholder: string;
-  value: string;
-  onChange: (value: string) => void;
-  size?: TextFieldSize;
-}
